Styling fixes and adding Powerset test to pytest
